fix: Fix support for 'IN'.
The regex for '"field" IN (?, ?)' queries was not correctly picking up the placeholder. Even if it had, the values were set in an array rather than as a single string, so they wouldn't have all been passed to the closure correctly.

The regex now correctly picks up IN queries, and the values for those queries are now passed through to the closure correctly as a single array.